fix(mcp): address PR #48 review comments - type consistency
Addresses review comments by @copilot-pull-request-reviewer on PR #48:

1. Temperature validation consistency (tools.rs:232, 103):
   - Changed validation at line 232 from '< 0.0' to '<= 0.0'
   - Now consistently rejects 0K (absolute zero) across all temperature checks
   - Rationale: 0K should not be a valid filter value in any context

2. System ID type consistency (types.rs:126, 145, 167, 179, 201):
   - Changed all system_id and destination_id fields from u32 to u64
   - Affects: Waypoint, SystemSummary, SystemDetails, NearbySystem, GateConnection
   - Matches specification in specs/016-mcp-server-integration/data-model.md
   - Aligns with SQLite INTEGER (i64) storage and prevents overflow for large IDs

Changes:
- tools.rs: Fixed temperature validation to reject 0K consistently
- types.rs: Updated 5 structs to use u64 for system identifiers

Validation:
- cargo clippy: clean (no warnings)
- cargo test -p evefrontier-mcp: 35/35 tests passing

Closes review threads:
- PRRT_kwDOQR89eM5nvke6 (temperature line 232)
- PRRT_kwDOQR89eM5nvke_ (temperature line 103)
- PRRT_kwDOQR89eM5nvkfB (system_id u32->u64)
- PRRT_kwDOQR89eM5nvkfE (system_id u32->u64)
- PRRT_kwDOQR89eM5nvkfI (system_id u32->u64)

+++ crates/evefrontier-mcp/src/types.rs
@@ -123,7 +123,7 @@ pub struct RouteDetails {
 #[derive(Debug, Clone, Serialize)]
 pub struct Waypoint {
     pub system_name: String,
-    pub system_id: u32,
+    pub system_id: u64,
     pub position: Position3D,
     pub min_temperature_k: f64,
     /// Type of jump to reach this system from previous
@@ -142,7 +142,7 @@ pub struct RouteError {
 #[derive(Debug, Clone, Serialize)]
 pub struct SystemSummary {
     pub name: String,
-    pub system_id: u32,
+    pub system_id: u64,
     pub position: Position3D,
     pub min_temperature_k: f64,
 }
@@ -164,7 +164,7 @@ pub struct SystemInfoOutput {
 
 #[derive(Debug, Clone, Serialize)]
 pub struct SystemDetails {
-    pub system_id: u32,
+    pub system_id: u64,
     pub name: String,
     pub position: Position3D,
     pub min_external_temperature_k: f64,
@@ -176,7 +176,7 @@ pub struct SystemDetails {
 #[derive(Debug, Clone, Serialize)]
 pub struct GateConnection {
     pub destination_system: String,
-    pub destination_id: u32,
+    pub destination_id: u64,
 }
 
 #[derive(Debug, Clone, Serialize)]
@@ -198,7 +198,7 @@ pub struct SystemsNearbyOutput {
 #[derive(Debug, Clone, Serialize)]
 pub struct NearbySystem {
     pub name: String,
-    pub system_id: u32,
+    pub system_id: u64,
     pub distance_ly: f64,
     pub min_temperature_k: f64,
 }
